requirements.txt
beautifulsoup4>=4.9.0
Pillow>=8.0.0
piexif>=1.1.0